--- include/epri.h
@@ -1,26 +1,0 @@
-/*	SCCS Id: @(#)epri.h	3.4	1997/05/01	*/
-/* Copyright (c) Izchak Miller, 1989.				  */
-/* NetHack may be freely redistributed.  See license for details. */
-
-#ifndef EPRI_H
-#define EPRI_H
-
-struct epri {
-	aligntyp shralign;	/* alignment of priest's shrine */
-				/* leave as first field to match emin */
-	schar shroom;		/* index in rooms */
-	coord shrpos;		/* position of shrine */
-	d_level shrlevel;	/* level (& dungeon) of shrine */
-	boolean pbanned;	/* player banned by priest */
-	char signspotted;	/* max number of signs spotted by priest */
-};
-
-#define EPRI(mon)	((struct epri *)&(mon)->mextra[0])
-
-/* A priest without ispriest is a roaming priest without a shrine, so
- * the fields (except shralign, which becomes only the priest alignment)
- * are available for reuse.
- */
-#define renegade shroom
-
-#endif /* EPRI_H */
