+++ src/pray.c
@@ -3,7 +3,7 @@
 /* NetHack may be freely redistributed.  See license for details. */
 
 #include "hack.h"
-#include "epri.h"
+#include "mextra.h"
 #include "artifact.h"
 #include "artilist.h"
