+++ src/read.c
@@ -3,7 +3,7 @@
 /* NetHack may be freely redistributed.  See license for details. */
 
 #include "hack.h"
-#include "edog.h"
+#include "mextra.h"
 #include "artifact.h"
 
 /* KMH -- Copied from pray.c; this really belongs in a header file */
