+++ src/wizard.c
@@ -9,7 +9,7 @@
 
 #include "hack.h"
 #include "qtext.h"
-#include "epri.h"
+#include "mextra.h"
 
 extern const int monstr[];
