+++ src/xhityhelpers.c
@@ -5,7 +5,7 @@
 #include "hack.h"
 #include "artifact.h"
 #include "monflag.h"
-#include "edog.h"
+#include "mextra.h"
 #include "xhity.h"
 
 extern boolean notonhead;
